Module sunrise_bootstrap::bootstrap_logging

bootstrap logging on rs232

A pale copy of the rs232 kernel device. Used by the bootstrap stage to provide some logging.

This driver is meant to be as simple as possible

Structs

Serial
A logger that sends its output to COM1.

Constants

COM1

Functions

bootstrap_inb⚠

bootstrap_log
Sends a string to COM1.

bootstrap_outb⚠

init_bootstrap_log⚠
Init the rs232 COM1. Must be called before logging anything.