Function sunrise_bootstrap::bootstrap_start

#[no_mangle]pub unsafe extern "C" fn bootstrap_start() -> !

The very start.

We are called from grub, with the address of the multiboot informations in $ebx.

What we do is :

bzero the .bss.

make $esp and $ebp point to STACK.

call do_bootstrap, passing it $ebx.

Safety

This may only be called once, as the bootstrap entrypoint.