Struct sunrise_bootstrap::frame_alloc::Frame

pub struct Frame {
    physical_addr: usize,
    is_allocated: bool,
}

A pointer to a physical frame

A frame is 4ko in size

Should only be used when paging is off

Fields
physical_addr: usizeis_allocated: bool
Implementations

impl Frame

pub fn address(&self) -> PhysicalAddress

Get the physical address of this Frame

pub(super) fn is_allocated(
&self
) -> bool

Gets the current allocation state

pub fn from_physical_addr(physical_addr: PhysicalAddress) -> Frame

Constructs a frame structure from a physical address

This does not guaranty that the frame can be written to, or even exists at all

Panic

Panics when the address is not framesize-aligned

pub unsafe fn from_allocated_addr(physical_addr: PhysicalAddress) -> Frame

Constructs a frame structure from a physical address

This does not guaranty that the frame can be written to, or even exists at all

Panic

Panics when the address is not framesize-aligned

Safety

This function should only be used on physical_addr that are reserved - that is, addresses that the frame allocator will never give. Otherwise, it is unsound.