Module sunrise_bootstrap::gdt

GDT Handler

The Global Descriptor Table is responsible for segmentation of memory. In our case though, we don't really care about that.

Modules

i386
This crate is x86_64's little brother. It provides i386 specific functions and data structures, and access to various system registers.

segment_selector
Types for the Global Descriptor Table and segment selectors.

Structs

DescriptorTable
A structure containing our GDT. We can have at most 16 segments, we should be more than enough.

DescriptorTableEntry

FAULT_TASK

MAIN_TASK

Enums

SystemDescriptorTypes

Statics

GDT
The GDT set-up by the bootstrap

GLOBAL_LDT
The global LDT used by all the processes.

Functions

init_gdt