Trait sunrise_bootstrap::paging::table::detail::I386PageTablesSet

pub trait I386PageTablesSet {
    type PageDirectoryType: PageDirectoryTrait;
    fn get_directory(
        &mut self
    ) -> SmartHierarchicalTable<Self::PageDirectoryType>;
}

A trait describing the interface of a PageTable hierarchy.

Implemented by ActivePageTables, InactivePageTables and PagingOffPageSet

Implementer only has to provide the type of the directory it will provide, and a function to map it.

Thanks to this we can have the same api for every kind of page tables, the only difference is the way we access the page directory :

an ActivePageDirectory will want to use recursive mapping

an InactivePageDirectory will want to temporarily map the directory

a PagingOffPageDirectory will point to physical memory

Associated Types

type PageDirectoryType: PageDirectoryTrait

Loading content...
Required methods

fn get_directory(&mut self) -> SmartHierarchicalTable<Self::PageDirectoryType>

Gets a reference to the directory

Loading content...