Trait sunrise_bootstrap::paging::table::HierarchicalTable

pub trait HierarchicalTable {
    fn entries(&self) -> &[Entry; 1024];

    fn entries_mut(&mut self) -> &mut [Entry; 1024];

    fn zero(&mut self) { ... }

    fn map_nth_entry<T: Flusher>(
        &mut self, 
        entry: usize, 
        frame: Frame, 
        flags: I386EntryFlags
    ) { ... }

    fn guard_nth_entry<T: Flusher>(&mut self, entry: usize) { ... }

    fn flush_cache() { ... }
}

A table of entries, either the directory or one of the page tables

Required methods

fn entries(&self) -> &[Entry; 1024]

fn entries_mut(&mut self) -> &mut [Entry; 1024]

Loading content...
Provided methods

fn zero(&mut self)

zero out the whole table

fn map_nth_entry<T: Flusher>(
&mut self,
entry: usize,
frame: Frame,
flags: I386EntryFlags
)

Creates a mapping on the nth entry of a table T is a flusher describing if we should flush the TLB or not

fn guard_nth_entry<T: Flusher>(&mut self, entry: usize)

Marks the nth entry as guard page T is a flusher describing if we should flush the TLB or not

fn flush_cache()

Loading content...