Trait sunrise_bootstrap::paging::table::PageTableTrait

pub trait PageTableTrait: HierarchicalTable {
    type FlusherType: Flusher;
    fn map_whole_table(
        &mut self, 
        start_address: PhysicalAddress, 
        flags: I386EntryFlags
    ) { ... }

    fn map_guard_whole_table(&mut self) { ... }
}

Associated Types

type FlusherType: Flusher

Loading content...
Provided methods

fn map_whole_table(
&mut self,
start_address: PhysicalAddress,
flags: I386EntryFlags
)

👎 Deprecated

Used at startup when creating the first page tables.

fn map_guard_whole_table(&mut self)

Used at startup when creating the first page tables.

Loading content...