Trait sunrise_bootstrap::paging::VirtualSpaceLand

pub trait VirtualSpaceLand {
    fn start_addr() -> VirtualAddress;

    fn end_addr() -> VirtualAddress;

    fn start_table() -> usize { ... }

    fn end_table() -> usize { ... }
}

A trait describing the splitting of virtual memory between Kernel and User. Implemented by UserLand and KernelLand

Required methods

fn start_addr() -> VirtualAddress

The first address in this land.

fn end_addr() -> VirtualAddress

The last address in this land.

Loading content...
Provided methods

fn start_table() -> usize

The index in page directory of the first table of this land

fn end_table() -> usize

The index in page directory of the last table of this land

Loading content...